Login Validation

+++ Customer_management_system/src/com/customer_management_system/ServicePlans.java
@@ -3,7 +3,7 @@ package com.customer_management_system;
 public enum ServicePlans {
     SILVER(1000),GOLD(2000),DIAMOND(5000),PLATINUM(10000);
 	
-	private int charges;
+	private double charges;
 
 	private ServicePlans(int charges) {
 		this.charges=charges;
@@ -14,7 +14,7 @@ public enum ServicePlans {
 		return name()+" : "+charges;
 	}
 	
-	public int getCharges() {
+	public double getCharges() {
 		return charges;
 	}
